requirements/requirements-dev.txt
djangorestframework
Pillow
drf-yasg
django-cors-headers
psycopg2-binary
django-filter
redis
djoser
djangorestframework-simplejwt
celery
python-dotenv
python-slugify

django-silk